[feat] 세션 별 고유 id 부여하는 기능 추가; 로직 서버 및 클라 접속 시 고유 세션 id를 부여 후 전달하도록 제작; 패킷 프로토콜과 타입 분리; 클라이언트 세션 개별적으로 I/O가 가능하도록 HashMap 사용해서 Session들을 관리하도록 변경

+++ IocpLib/Fundamentals/Enumerations.h
@@ -1,7 +1,7 @@
 #pragma once
 #include "../pch/pch.h"
 
-enum HostTypeEnum : uint8
+enum HostTypeEnum : uint16
 {
 	NONE,
 	CLIENT_HEADLESS,
